Relocate the JSON based result models
