+++ metasource/models/dict/misc.go
@@ -1,13 +1,9 @@
-package json
+package dict
 
 type Version struct {
 	Epoch   string `json:"epoch"`
 	Version string `json:"version"`
 	Release string `json:"release"`
-}
-
-type Repo struct {
-	Repo string `json:"repo"`
 }
 
 type UnitBase struct {
